Move ScanManagement in a dedicated user control to move all its code out of the main window Create Helpers folder for helper scripts and move Extensions.cs in it Create DesignerMode helper script to prevent compilation error in user controls constructor
