+++ ScanNetDownloader/Logic/FileManagement.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Win32;
 using ScanNetDownloader.View.CustomControls;
+using ScanNetDownloader.Logic.Helpers;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
