+++ ScanNetDownloader/MainWindow.xaml
@@ -7,10 +7,6 @@
         xmlns:customControls="clr-namespace:ScanNetDownloader.View.CustomControls"
         mc:Ignorable="d"
         Title="ScanNetDownloader" Height="720" Width="1280">
-    <Window.Resources>
-        <CollectionViewSource x:Key="ScanItemsSource" Source="{Binding ScanListItems}">
-        </CollectionViewSource>
-    </Window.Resources>
     <Grid x:Name="gridMainContent">
 
         <Grid.RowDefinitions>
@@ -23,89 +19,7 @@
             
             <!--Scan Management-->
             <TabItem Name="tabScanManager" Header="Scan manager">
-                <Grid>
-                    <Grid.RowDefinitions>
-                        <RowDefinition/>
-                        <RowDefinition Height="50"/>
-                    </Grid.RowDefinitions>
-
-
-                    <Border Margin="10 10 10 0">
-                        <!--<DataGrid Name="listVwScans" AutoGenerateColumns="False" CanUserAddRows="False" CanUserDeleteRows="False" ItemsSource="{Binding Source={StaticResource ScanItemsSource}, UpdateSourceTrigger=PropertyChanged}">
-                            <DataGrid.Columns>
-                                <DataGridCheckBoxColumn Header="Download" Binding="{Binding Path=IsSelectedForDownload}"/>
-                                <DataGridTextColumn Header="BookName" Binding="{Binding Path=BookName}"/>
-                                <DataGridTextColumn Header="Chapter" Binding="{Binding Path=ChapterId}"/>
-                                <DataGridTextColumn Header="Number of Pages" Binding="{Binding Path=PagesCount}"/>
-                                <DataGridTextColumn Header="Website" Binding="{Binding Path=Website}"/>
-                                <DataGridTemplateColumn Header="Options">
-                                    <DataGridTemplateColumn.CellTemplate>
-                                        <DataTemplate>
-                                            <TextBlock Text="{Binding}"/>
-                                            --><!--<ListView ItemsSource="{Binding}"/>--><!--
-                                        </DataTemplate>
-                                    </DataGridTemplateColumn.CellTemplate>
-                                </DataGridTemplateColumn>
-                            </DataGrid.Columns>
-                        </DataGrid>-->
-                        <ListView Name="listVwScans" ItemsSource="{Binding Source={StaticResource ScanItemsSource}, UpdateSourceTrigger=PropertyChanged}" SelectionMode="Single" HorizontalContentAlignment="Stretch" Background="LightGray">
-                            <ListView.ItemContainerStyle>
-                                <Style TargetType="ListViewItem">
-                                    <Setter Property="Background" Value="White"/>
-                                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
-                                </Style>
-                            </ListView.ItemContainerStyle>
-                            <ListView.View>
-                                <GridView AllowsColumnReorder="True">
-                                    <GridViewColumn Header="Download" Width="60">
-                                        <GridViewColumn.CellTemplate>
-                                            <DataTemplate>
-                                                <CheckBox IsChecked="{Binding Path=IsSelectedForDownload}" ToolTip="Select this scan for download" HorizontalAlignment="Center"/>
-                                            </DataTemplate>
-                                        </GridViewColumn.CellTemplate>
-                                    </GridViewColumn>
-                                    <GridViewColumn Header="Book Name" DisplayMemberBinding="{Binding Path=BookName}" Width="450"/>
-                                    <GridViewColumn Header="Chapter" Width="100">
-                                        <GridViewColumn.CellTemplate>
-                                            <DataTemplate>
-                                                <TextBlock Text="{Binding Path=ChapterId}" TextAlignment="Center" VerticalAlignment="Center"/>
-                                            </DataTemplate>
-                                        </GridViewColumn.CellTemplate> 
-                                    </GridViewColumn>
-                                    <GridViewColumn Header="Number of pages" Width="100">
-                                        <GridViewColumn.CellTemplate>
-                                            <DataTemplate>
-                                                <TextBlock Text="{Binding Path=PagesCount}" TextAlignment="Center" VerticalAlignment="Center"/>
-                                            </DataTemplate>
-                                        </GridViewColumn.CellTemplate>
-                                    </GridViewColumn>
-                                    <GridViewColumn Header="Website" Width="250">
-                                        <GridViewColumn.CellTemplate>
-                                            <DataTemplate>
-                                                <TextBlock Text="{Binding Path=Website}" TextAlignment="Center" VerticalAlignment="Center"/>
-                                            </DataTemplate>
-                                        </GridViewColumn.CellTemplate>
-                                    </GridViewColumn>
-                                    <GridViewColumn Header="Options" Width="250">
-                                        <!--I would love to understand why scanItem is displayed in this column and how to do it with a cell template to make it more clear
-                                        but I can't find how to assign the source scanItem as the template scanItem-->
-                                    </GridViewColumn>
-                                    <!--<GridViewColumn Header="Url" DisplayMemberBinding="{Binding Path=Url}"/>-->
-                                </GridView>
-                            </ListView.View>
-                        </ListView>
-                    </Border>
-                    <StackPanel Name="stPanelNoScans" VerticalAlignment="Center">
-                        <TextBlock Text="No scans added yet. Add a new scan url to get the scan information and download it." VerticalAlignment="Center" HorizontalAlignment="Center" Margin="0 20 0 10"/>
-                        <Button Content="＋" Click="btnAdd_Click" Width="50" Height="50" FontSize="25"/>
-                    </StackPanel>
-                    
-
-                    <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Center">
-                        <Button Name="btnAdd" Content="Add" Click="btnAdd_Click" Height="40" Width="100" Margin="5"/>
-                        <Button Name="btnStart" Content="Start" Click="btnStart_Click" Height="40" Width="100" Margin="5"/>
-                    </StackPanel>
-                </Grid>
+                <customControls:ScanManagerView x:Name="scanManagerVw" StartDownload="scanManagerVw_StartDownload"/>
             </TabItem>
             
             <!--Scan Download-->
